Redirect the User to Last Page Upon Login
This change will place the user back at the last page they had visitied when the log in. For example, if the user visits /compose and aren't logged in, upon logging in they will be redirected back to /compose. The particualr applicaiton of this is when a user that isn't logged in tries to open a Data One package and needs to log in.