Fix Missing Tales
My autoformatter changed a few lines, but the ones of interest are in the `searchFilter` function on line 195. I wasn't able to reproduce the problem on the new ui, but I was able to under the master branch. I tested the code under master (and it fixed the issue), and I wasn't able to see any difference in the newui branch. In each failure case, `searchView` was empty at the time
`component.set` and `component.paginate` were called.