Set the Filter on Init
`searchFilter` gets called when the user enters a search term. It searches through the `filteredSet` object for tale titles that match the user supplied text.

The `filteredSet` object is sometimnes empty, which causes the search to fail. The main cause was not setting the `filteredSet` object when the widget is initialized. This resuts in `filteredSet` never being set, and thus the failed searches.