Remove `Results` JQuery References
In component.js I removed the jquery references to the results section and added a variable, showResults, that is true/false when we want to show/hide the div.

In template.hbs I added an if statement around the results div. There are some formatting changes in between the if openeing and if closing statements (which can be ignored).