Refactor to Open OCRID Login in Run Page
This commit adds the paddleboard at the top of the run page with some additional behavior.

If the user hasn't authenticated with DataONE and they click the publish button, another popup will open letting them know they're about to leave Whole Tale. If they click yes, they're sent through the DataONE-ORCID auth. They are redirected to `/run/taleID?auth=true`. If they hit Cancel they aren't.

If they click Publish when they have already logged in, the modal should pop open.

The queryParam `auth` needs to be checked in the component when the page is loaded. If it's true, the publish dialog should open.