Populate the Compose page on visit from dataverse.
This commit extends the component to handle users coming from dataverse.

Because dataverse & dataone nodes can place their api endpoints anywehre, I added a query parameter to the route that takes the node's API endpoint. This way, the query for the package name can be appended to end.

`renderDataversePackage` is the portion that asks dataverse about the package. The appropriate dataverse api call is `/datasets/:persistentId?persistentId=`. The Tale name is set by the name of the package., The `Input data` field is set by setting `this.inputData` to the package name.

I scheduled a function that checks if we're coming from a third party, and if so, decide whether to call the DataONE (`renderDataONEPackage`) or dataverse (`renderDataversePackage`) communication code.

The `Input data` section in the UI is a rendering of the Array of Objects, `this.inputData`. The objects stored in `this.inputData` need to have a `name` attribute. Each time a user selects a file in the File Explorer, we reset `this.inputData` to the selected files. Because our dataset isn't registered yet, it won't be in the file system, and thus not in `this.inputData`.

`insertPackageName` is the function that handles inserting the package name into `this.inputData`. It includes a check to see that the data_api query parameter exists.