Remove Query to Third Parties & Take Package Title as a Query Parameter
We shouldn't have to query the referring site for information about the package (we're only after the name at this point). Instead, we can just take it as a query parmeter and use it for the Tale Title and data folder name.

Since we don't need to know the difference between dataverse and dataone, I renamed `data_provider` to `data_title`. `data_title` holds the title of the package.

I removed the check for the third party in the scheduled method (line 62). I removed `renderDataONEPackage` and renamed `renderDataversePackage` to `addExternalPackage`.