Overhaul Publishing Modal
I added an accordion dropdown to the publishing modal that shows the Tale's publishing status. This includes a progress bar and other info such as the doi and package location.

In the pubblishing component, I moved a number of component variable initializations into the `init()` method. I also added a check in there to see if the Tale
   a. Has been published already
   b. Is being published right now
I also renamed the license and metadataimmuteable files.

A little re-working was done on getting files out of the file browser, which was done in `prepareItemIds` and `getItems`.

`handlePublishingStatus` Is the heart of updating the progress bar and letting the user know the status of publication. It polls the job system until it gets an error or success status.