Hide the Edit Tale button and restrict publishing to the Tale owner
The 'Edit Tale' modal isn't complete yet, so we need to hide the button that activates it. I did this by commenting it out.

I also conditionally restricted the publishing button.
After rendering, I added a check to see if the Tale's creator ID matches the logged in user's ID. If so, the button enables.

I also removed `getDataONEJWT` because it was moved into a service, and the service is used instead.